Added Code Examples page
This page is a general base of code examples that can be used by veterans or new users to developing add-ons for XenForo. The page is in no way complete as there will always be developers adding to this page of various PHP code examples as well as template code examples.